+++ src/main/java/com/example/car/management/filters/MaintenanceMonthlyReportFilter.java
@@ -4,24 +4,24 @@ import java.time.LocalDate;
 
 public class MaintenanceMonthlyReportFilter {
 
-    private Integer garageId;
+    private Long garageId;
     private LocalDate startMonth;
     private LocalDate endMonth;
 
     public MaintenanceMonthlyReportFilter() {
     }
 
-    public MaintenanceMonthlyReportFilter(Integer garageId, LocalDate startMonth, LocalDate endMonth) {
+    public MaintenanceMonthlyReportFilter(Long garageId, LocalDate startMonth, LocalDate endMonth) {
         this.garageId = garageId;
         this.startMonth = startMonth;
         this.endMonth = endMonth;
     }
     
-    public Integer getGarageId() {
+    public Long getGarageId() {
         return garageId;
     }
 
-    public void setGarageId(Integer garageId) {
+    public void setGarageId(Long garageId) {
         this.garageId = garageId;
     }
